// src/profile/entity/profile.enum.ts
export enum ProfileColumns {
    ID = 'id',
    PROFILE_IMAGE = 'profile_image',
    FIRST_NAME = 'first_name',
    LAST_NAME = 'last_name',
    BIO = 'bio',
    PHONE_NUMBER = 'phone_number',
    LINKEDIN = 'linkedin',
    GITHUB = 'github',
    TWITTER = 'twitter',
    CREATED_AT = 'created_at',
    UPDATED_AT = 'updated_at',
}

export enum ProfileRelations {
    ACCOUNT = 'account',
}